+++ org.jebtk.modern/src/main/java/org/jebtk/modern/css/CSSNumProp.java
@@ -1,9 +1,9 @@
 package org.jebtk.modern.css;
 
 public interface CSSNumProp extends CSSProp {
-  
+
   public int getInt();
-  
-  public double getDouble();
+
+  public double getFloat();
 
 }
